Update for 4.2-0 (v2018.2)
Includes update of translation files, ffho-packages for
WiFi-autoupdater-fallback and an AP timer.
This version should be much more stable as it has less memory
consumption than v2017.x.
Also changed channel and BSSID settings back as we have not yet
switched to a firmware where all nodes have the autoupdater-
fallback. Next step after the network is stable again must be to
introduce domain-specific settings and use the domain-switcher
package to split the network.